Rag/requirements.txt
langchain==0.2.11
langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-community==0.2.10
pypdf==4.3.1
wikipedia==1.4.0
#langchain-chroma==0.1.2
faiss-cpu==1.8.0.post1
bs4==0.0.2
unstructured==0.15.0
langchain_huggingface
langchain-ollama == 0.1.3
pymilvus ==2.5.0
langchain_milvus ==0.1.8
neo4j==5.26.0
langchain-groq==0.1.6
langchainhub
langserve==0.2.2
langgraph==0.1.19